backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.5
google-auth==2.40.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-cloud-tasks==2.19.3
firebase-admin==7.1.0
mistralai==1.9.10
python-multipart
reportlab==4.2.5
